Replace SearchBarView with SubSearchBar
SubSearchbar is built according to Elm architecture style.

This also seems to have resolved a puzzling interaction bug I was having
previously, where the "<" button was not triggering a return to the
stream... but only after interacting with other elements, such as the
editor. Anyway, here's hoping it was just a logic bug in the old
bindings. The Elm Architecture makes it so much easier to avoid these
related state bugs in the first place, and fix them when they do appear.