Move app model up, treat ContentView as equatable
This prevents ContentView from re-rendering if AppModel is the same.

Right now, this doesn't change much, but in future we can bring some
service-related state up to the app level, and move ContentView-related
state into a content field of the AppModel. This will allow us to use
Elm-style update functions to manage the lifecycle of things like search
indexes.

It does mean a static preview for ContentView, but that's ok. We're
getting to the point where you want to do a full build to test anyway.