Fix thread expansion
Bug was caused by a mutability footgun.

Previously, I was using `Array.first` to get the first thread, but because
Swift copies value types by default, this was giving me a copy. I would then
pass that copy down to be mutated, and it would be, but the copy was an
orphan, so it resulted in no visible change to interface.

To avoid this, should consider refactor where we return references to new
model instead of mutating model in place.