Attach delegate, and remove focus
Focus `becomeFirstResponder()`, `resignFirstResponder`, was causing a cycle
problem in SwiftUI. I think because we render and then immediately
re-render the view? Anyway, calling these methods in an async block on
the main thread will fix the cycle problem. However, there are some
complex state-related issues around handling focus in the model. For the
moment, I've just removed it.