Save file modified time as Unix Timestamp seconds
Modified time on file, stored as Unix Timestamp Integer
(rounded to the nearest second).

We were previously getting what appeared to be rounding precision errors
when serializing datetimes as ISO strings using .

Additionally, file timestamps precision is limited to:
1 second for EXT3
1 microsecond for UFS
1 nanosecond for EXT4

To-the-nearest-second precision is fine for the purpose of comparing changes,
and handwaves away these issues.