Have editor handle its own IO
Slowly porting everything away from sending actions through DB
component. Instead, we'll have a DatabaseService, and pass that down to
update functions, or else pass down facades.

The biggest win here is that we can keep the database service to a minimum set
of methods, and we don't need to proliferate actions for specific modules.

This also gets rid of the problem of action ambiguity. To properly have the
database component serve all of its siblings, we would have had to
create parallel actions in DB that matched all of the IO actions in the
sibling modules. We already had an ambiguous case with readEntry which
was used to do something more specific, but that semantics was lost as
we routed IO actions around from sibling to Database component and back.

By using the service directly, the module can use task-appropriate actions,
and we don't need a Database component to embody the complexity of all of the
actions of all of the sibling components it serves.