Remove main thread scheduler from send
This makes send thread UNsafe, but it is for a good cause: SwiftUI
transactions are required to be syncronous, and when we scheduled
actions on the main thread, it causes them to get bumped to the next
tick. To play nicely with transactions, withTransaction, and
withAnimation, you MUST have a main thread syncronous send method.