Remove animation to avoid state-related bug with focus
Focus is bi-directional. It can be set from the outside, and also via
event listeners didBeginEditing, didEndEditing. Because of this, we get
small cycles where you set focus by sending setFocus, which triggers
didBeginEditing, which sends a setFocus. Adding an animation to the mix
causes further issues. Removing the animation is an ok workaround for
now.