Update Store to use a unified reducer
The distinction between update reducer and effect producer introduced
complications for effects that rely on state. Since some effects might
require previous state, and other effects might require next state, the result
was that you ended up having to derive state in both effect and update,
leading to duplication of logic.

The new approach returns to a non-mutating update function that is
responsible for returning a tuple of (state, effect).

This has the added benefit of only mutating state once for each update,
reducing the number of times we publish objectWillChange. Previously, we
used mutation, under the theory that it would be cheaper. Of course,
each value-type mutation fires objectWillChange once. Without
understanding the internals of SwiftUI, it is doubtful that this
triggers over-rendering. However, it does feel unnecessary, and the
clarity of immutable updates are nice anyway.