Introduce title + excerpt EntryStub
This is more efficient than holding entire subtext files in memory for
lists. We read in the content, parse it, then produce the excerpt,
throwing away the parsed content.

The model that was previously called Stub is now EntryLink (it's a title
plus a slug), and title + excerpt + slug is called EntryStub.