Reverse chron note list (#38)
Fixes https://github.com/gordonbrander/subconscious/issues/12.

This PR also "un-themes" the background of a number of components. See https://github.com/gordonbrander/subconscious/issues/39.

* Set higher limit for notes

* Basic proof-of-concept reverse-chron list

* Wire up list item tap to load

* Set background to white

* Remove background set

* Remove background set

* Customize NavigationBar title via toolbar

Accept the limitations of SwiftUI and just do a custom toolbar without
the big text.

* Set vertical padding on list items

* Introduce title + excerpt EntryStub

This is more efficient than holding entire subtext files in memory for
lists. We read in the content, parse it, then produce the excerpt,
throwing away the parsed content.

The model that was previously called Stub is now EntryLink (it's a title
plus a slug), and title + excerpt + slug is called EntryStub.