Add environment to store (#95)
Fixes https://github.com/gordonbrander/subconscious/issues/89
Fixes https://github.com/gordonbrander/subconscious/issues/61

## Motivation

How to integrate long-lived services into Store/SwiftUI? (See https://github.com/gordonbrander/subconscious/wiki/swiftui-services)

We had previously created an `init` function in the root `App`, where we constructed the store, as well as services which produced publishers. This allowed us to pass the publisher from services into the `subscription` field of Store.

However, this felt weird and created a strange bifurcation between `DatabaseService` and `KeyboardService`, where `DatabaseService` was a global constant and `KeyboardService` was instantiated in `App` `init`.

Furthermore, it seems that Apple's docs suggest hanging root state off of `App`, in an `@StateObject` (https://developer.apple.com/documentation/swiftui/app). Stands to reason. The thing is, `@StateObject` does not allow assignment in `init`. You may only assign directly as part of the property wrapper definition.

So our current method of subscribing to long-lived services runs against the grain of SwiftUI APIs.

## Solution

Introduce an `Environment` to Store. `Environment` is a place to put APIs, services, and other information. It gets passed to `update`, which may use it to derive `Change`s.

`Environment` is a perfect place to put long-lived services, such as keyboard listeners. Its lifetime matches the lifetime of the app. Furthermore, we can subscribe to asynchronous from services the same way we do with any other method call: by mapping a publisher to an `fx` publisher, and then returning it as part of a `Change`. Store will hang on to the cancellable until it completes. In the case of a long-lived service, it will only complete when the app is killed.

We use our app-level `.appear` event as an opportunity to subscribe to long-lived effects. This is how we subscribe to keyboard events now.

This also allows for more dynamism in the way that services are constructed. It becomes possible to hot-swap services, etc.