Only set state on store if it has changed
This prevents unnecessary body recomputes in views.

This is especially valuable now that we are polling every 15s. No need
to recompute body if nothing changes in state.

This means we require Equatable from now on. It's a good idea anyway,
since it frees us up to use EquatableView wherever we please.