Autosave (#112)
* Add periodic poll service to environment

* Kick off polling service in appear

Log action for now

* Autosave current entry periodically

Saves current entry, if any, as scheduled by our poll service.

* Disambiguate save, autosave, done

* Set implicit animation for changes to recents

Set transition as well, as it prevents weird initial states.

* Only set state on store if it has changed

This prevents unnecessary body recomputes in views.

This is especially valuable now that we are polling every 15s. No need
to recompute body if nothing changes in state.

This means we require Equatable from now on. It's a good idea anyway,
since it frees us up to use EquatableView wherever we please.

* Do not show recents when zero elements

This causes transition (opacity) to play instead of animation
(grow/shrink). Important since the results may appear after onAppear,
given the fetch effect is asyncronous.

* Allow marking saved state when setting editor

This keeps us from autosaving when we don't need to.

- Save when exiting detail view
- Avoid triggering save if nothing to save
- Check for save state before saving (i.e. save is no longer force-save)

* Use updateEditorDom instead of setEdiorDom for back

* Add enum for modeling async save state

* Model save state as enum