Fade in transition when clicking link (#202)
Implementation:
UIViewRepresentable seems unable to play nicely with SwiftUI animations.
Given this, we instead overlay a white rect, show it immediately when
clicking a link, then fade it out.

We also introduce "isDetailLoading" flag that allows us to show loading
scrim without resetting editor. This prevents a reflow of the contents
of the editor, and avoids the visual tearing that we were getting.

The result is a smooth animated transition between detail states.