Add Unit Test target (#203)
...and stub out Slug tests.

It turns out that the tests which XCode stubs out for you in the XCode
SwiftUI project builder are UI Tests, which are a completely different
kind of target than Unit Tests. To write Unit Tests you must create a
new Unit Test target in the project definition.

I've done that here. Added a SubconsciousTests target which is a Unit
Tests target.