TextStorage markup renderer (#220)
Replace renderer with TextStorage-based markup renderer. TextStorage is the lower-level API by which Apple seems to want us to render markup.

Fixes #211.

## Details

In the past, we would use a UITextViewDelegate to wait for text changes to occur, then look at the text, construct a Subtext DOM, render out a new NSAttributedString, and set that string on the UITextView. This effectively bashed the UITextView text, updating it twice for every change. My posit at the time was that since this all happens in the same tick, it shouldn't really cause any problems. However, it seems that iOS does some stuff when you change the text, such as autocompletion, autocorrection, and this approach creates problems. This is the mechanism by which I posit we're hitting #211.

So! It turns out there is a lower-level API called `NSTextStorage`. Apple seems to want us to use that render markup. When you instantiate a `UITextView` it creates a vanilla `NSTextStorage` object under the hood. This object is responsible for doing all styling work, and has various lifecycle hooks for updating attributes in response to text changes.

The new renderer uses `NSTextStorageDelegate`, to hook into `willProcessEditing`. This happens just after the UITextView has updated its backing TextStorage, but before the OS has done font substitution, or notified the view to update. We use this hook to calculate a DOM for the string value of `NSTextStorage`, and then set attributes on `NSTextStorage` based on that DOM. Because `NSTextStorage` is a subclass of `NSMutableAttributedString`, it isn't ever replaced, just updated.

See https://github.com/gordonbrander/subconscious/wiki/TextKit for more background notes.

## Notes

- `Subtext.render` is replaced by static function `Subtext.renderAttributesOf`, whos job is to read the contents of a NSMutableAttributedString to a Subtext Dom, and then set appropriate attributes on that NSMutableAttributedString.
- `MarkupTextViewRepresentable` takes an escaping closure `renderAttributesOf` who's job is to do the rendering.
- `MarkupTextViewRepresentable` now consumes a string binding. No style properties are exposed outside of the View. The View is expected to handle all attribute rendering.
- `Coordinator` now conforms to `NSTextStorageDelegate` and is set as the `UITextView.textStorage.delegate` and handles rendering the markup at the appropriate time