Check field before setting focus nil (#243)
Introduce `field` field to setFocus. When focus is resigned, we check
that the current focus state is actually the field that is expected
before setting nil. If it is some other state, we leave it alone.

In general nil focus means "relinquish my focus", not "relinquish all
focus". Since focus APIs are async, this distinction matters. If we're
trying to resign focus and someone else has taken focus in the meantime,
then we consider it job well done, and leave the state alone. This
prevents race conditions where the focus state gets stomped.

This PR also removes `resignFocus` in favor of using `setFocus` everywhere, since now `setFocus` has essentially the same behavior as `resignFocus` in cases where the focus being set is nil.

Fixes https://github.com/gordonbrander/subconscious/issues/242