Fix renaming deep slugs (#264)
Fixes https://github.com/gordonbrander/subconscious/issues/263

Previously we used file manager's removeEntry method to move the file,
but this is unable to create interim directories where needed.

This new approach uses the same logic as creating a new note. It first
reads the contents of the old note, then creates a new note with the old
contents at the new locations, writes it, then deletes the old one.