Introduce syntax for wikilinks, bold, italic, code (#262)
Part of #205 

This PR introduces the following markup to Subtext:

```
[[Wikilinks]]
*bold*
_italic_
`code`
```

Wikilinks are normalized to slugs, such that `[[A wikilink]]` becomes `/a-wikilink` under the hood. You can think of them as an alternative syntax for slashlinks.

I'm also opting for the Slack `_italic_` and `*bold*` markup over Markdown's goofy choice. This is a departure from Markdown, but I think a worthy one. We can reconsider before launch.

## Prior art

### Markdown

```
**bold**
*italic*
__bold__
_italic_
`code`
```

### Slack

```
*bold*
_italic_
`code`
```

### Roam

```
**bold**
__italic__
`code`
```

### Discord

```
**bold**
*italic*
_italic_
__underline__
`code`
```