Use title for wikilinks when slugified title matches valid slug (#276)
Uses title from entry for wikilinks, when appropriate. Falls back to sentence-ified slug when slugified title does not match slug.

- Replaces `EntryWikilink` struct with `EntryLink` struct. `EntryLink` allows title to diverge from slug.
- Introduces `Markup.Wikilink(_ link: EntryLink)`, which can derive a valid wikilink from the entry link.
  - To do this, we slugify the entry link title, and then compare it against the entry link slug. If the slugified title matches the slug, then we use the title as the wikilink text. If it does not match, then we sentence-ify the slug instead.

This allows us to default to nice-names when inserting wikilinks in the typical case where title can be converted to slug for article. Nice-names = original case and special characters are preserved.