Fix backtrack infinite loop in slashlink parse
We mistakenly called `tape.start()` instead of `tape.save()` in
`consumeInlineWordBoundaryForm`. This caused an infinite loop when the
slashlink was placed at the beginning of the line, because backtracking
would jump to the beginning of the line, then go again, etc.

This PR fixes the mistake and also adds regression tests.

Followup to https://github.com/gordonbrander/subconscious/pull/294