Accept unicode in URL and slashlink bodies (#302)
* Accept unicode in URL and slashlink bodies

We don't need to validate within the markup itself. Instead, we should
consider being tolerant of unicode characters, and santizing values from
markup before using them.

* Parse URLs before marking them up

* Only render wikilink bracket color if linked

It is possible for url closure to return nil. In that case, do not
mark up wikilink at all.

Previously we were marking up bracket range with light grey, but not
linking in this error case.