Save when navigating back from detail view (#316)
Fixes #308. Tested in simulator, and on device.

This PR accomplishes this primarily by introducing `setEditorFocus`, a specialized action that invokes the same logic as `setFocus`, but also invokes a save if the focus is not on the editor.

Adds test for `setEditorFocus` action.

Also refactors `DetailView` to use a `ViewStore`, a localized projection of Store for that view. When I set out to fix this bug, I thought a binding-like abstraction that allowed for `get/send` rather than `get/set` would be crucial. It turned out not to be necessary. However the refactor does improve the code, so I'm going to keep it.