Create shared static config instance
We need feature flags config statically available in too many places.
We should start with this assumption, rather than putting config in the model.

It may end up being the case later that we will want non-statically
available config that the user can modify in the view. This should
probably be done through UserDefaults.