Beginning focus refactor
Learnings: SwiftUI focus APIs are meant to be view-local. That is, there
is no global indicator of who is focused. The enum variant of focus APIs
is meant for single views with multiple controls.

Therefore: just manage focus as a local state of the view. Don't try to
store a global focus.