Request editor focus change when seting and presenting detail (#339)
Request editor focus change when seting and presenting detail.

The bug was caused by making the mistake of sending down an editorFocusChange event, which is a notification-style event of user-input-driven focus change. It does not schedule focus change. The fix sends down requestFocus, which does schedule focus change.

Fixes #338 and adds regression test.