Allow custom fallback content (#341)
This PR updates search-or-create actions to allow for the customization of fallback content that is distinct from the title of the note.

Previously we were using title as the fallback always, so these actions/APIs took a slug as a reference to the entry. This PR updates these actions/APIs to take an `EntryLink` instead, describing both a slug and a title. In some cases we continue to use entry title as fallback. However, adding this additional data to the APIs allows us to specify a title AND custom content if we wish.

This PR enables the work being done in https://github.com/subconsciousnetwork/subconscious/pull/340.