Persist notes to Noosphere (#393)
Naively read from and write to Noosphere when `Config.noosphere.enabled` is toggled on.

The Noosphere integration is not yet ready for user-facing use. This PR will landing with the config flag toggled off.

## Notes

I've refactored the data layer of the app to disentangle the database from the source-of-truth store. This made it possible for me to hide Noosphere store behind a config flag. The new persistence services look like this:

- `DataService`: the high-level service the store update method interfaces with. Generally returns Combine Publishers. It has references to:
    - `database: DatabaseService`.  a wrapper around SQLite3Database that has methods for queries, sync, and migrations
    - `memos: StoreProtocol`, a service persists to the file system (currently persisting header subtext).
    - `noosphere: NoosphereService`, an interface to Noosphere/the user's sphere.

Note that many features of the app still don't work right when Noosphere is toggled on. This is mostly because we're missing methods for sync and listing memos.