Major app tabs have their own store
Each major app tab manages its own store, with a small app store that
contains app-wide config shared between tabs.

This reduces app-wide state recalculations.

Tabs in iOS apps are not supposed to navigate between each other.
They are essentially meant to be separate mini-apps that happen
to exist in the same app. As such, they are a natural place to
split up app state, because there is minimal state shared
between tabs.

Dividing up the app this way optimizes performance, since changes in one
tab will not create view recalculations in the rest of the app. It is
also a bit easier to reason about.