Update to new version of ObservableStore
Deprecates ViewStore in favor of Address.forward. This is to resolve
shortcomings with the ViewStore approach when dealing with list items.
The live getter actually causes problems with list items, since you
might get a record that no longer exists.

This new approach resolves the issue by passing down bare state along
with a send function. Additionally, we now have a good way of dealing
with parent-child communication that is more complex than bindings for
views that also keep internal state.