Update Noosphere package (#404)
This picks up changes from Noosphere@v0.6.3 which fix the saving/rollback bug we
encountered in #397.

The v0.6.3 release branch Package.swift was incorrect, and points to the
v0.6.2 build artifact. Therefore, we stay on Noosphere@main, which picks
up the v0.6.3 build artifact.

Tested and confirmed that it resolves the saving problem.