Make MemoAddress an enum (#415)
Previously MemoAddress was a struct containing an enum. This PR changes it to be an enum.

Rationale: As we shift to multiplayer, we will need to introduce the concept of fully-qualified slashlinks (`@gordon/foo`) in addition to relative slashlinks (`/foo`)

However, a local address should always be a slug, never be fully-qualified, while a public address may be relative or fully qualified. This means local and public addresses will eventually need different data types. Turning MemoAddress into an enum allows us to support this branching logic.