Multiplayer @slashlink integration (#419)
This PR integrates `Slashlink` with `MemoAddress.public`.

Slashlink is a type that represents a `@petname/slug` or `/slug` slashlink type. Updating `MemoAddress` in this way allows us to address memos that belong to other spheres.

Prep work for #385. While this PR integrates the `Slashlink` type throughout the app, it does not expose `@petname/foo` markup in the editor yet. We will save that for a followup PR.

## Key changes

- Update `Slashlink` type for parsing correctness, and add numerous tests.
    - Validate via Regular Expression for improved correctness
    - Allow unicode word characters
- Update `Slug` type
    - Validate via Regular Expression for improved correctness
    - Allow unicode word characters
- Introduce `Petname` type, which represents a valid petname/nickname.
- Update `MemoAddress`
    - Public now holds a slashlink e.g. `MemoAddress.public(Slashlink)`
    - `MemoAddress.description` serializes as `public::@foo/bar`, `public::/bar`, `local::/bar`, etc.
- Allow reading from `@petname/foo` addresses for public memos