Render Subtext in SubtextTextParagraph
This allows us to hold on to the Subtext dom in the paragraph instance,
so we can inspect e.g. how many slashlinks there are in a paragraph.

Markup is only rendered for blocks that change. It appears that due to an
implementation detail in TextKit2 this will re-render the current block,
plus the adjacent block beneath. That's very effecient.