Rename isValid -> hasError
This allows for clearer of expression between the different layers of concern. The textfield should not show an error if it has not been touched yet, even if the underlying value is invalid. It's incorrect to say that the field is valid before it's been touched, but it also does not have an error.

Flipping the logic keeps the ValidatedTextField dumb and lets FormField expose both isValid and hasError properties.