Handle 3p slashlinks (#442)
We parse `@handle/links` but we need to actually do something with them
in the app. This PR threads that logic throughout the app so that we
open viewer detail views for 3p sphere content, and editor detail views
for our own content.

This PR threads 3p slashlink logic throughout the app, but stops short
of actually implementing the 3p content viewer, since that will take FFI
integration.

Slashlink handling is currently behind `Config.memoViewerDetailEnabled`
and flagged off, so clicking a qualified slashlink will do nothing. The
3p sphere content viewer is also mostly an empty view at this point, and
hidden behind the flag.

I would like to get this merged, then follow up with a PR for 3P content
viewer, and remove the flag.

Changes:

- Add support for editor views (own content) and viewer views (other's
content) to Navigation Stack. (Config flagged off.)
- Factored out Subtext renderer into separate struct.
  - Delegates for customizing link rendering
  - Disambiguate slashlink from wikilink parsing
- Changed `sub://` internal URL scheme to allow for `@` character.
    - Slashlink is now encoded as a query param.