Replace regex parser with recursive descent
This commit replaces the Regex-based matches with the old hand-rolled
recursive descent logic from
https://github.com/subconsciousnetwork/subconscious/commit/16709dfa95a99d37c8ce53423e2f1a9e4c3a7c5a.

All tests pass, except we now accept `@bare-handles-without-paths`
because it would be a bit of a pain to do backtracking. Not to worry,
the petname parsing code for URLs rejects bare petnames right now, so
the user-facing behavior is the same. And we'll want to support
petname-only slashlinks at some point anyway.