Implement basic 3p sphere content view (#486)
Progress toward #451.

This PR lands a basic 3P content viewer we can use to traverse the
Noosphere.

I ran into some minor problems with `Text` and `AttributedString` not
rendering links correctly. Rather than debug, I stubbed in a read-only
version of `SubtextTextViewRepresentable`. Additionally, the view has
basic handling for loading and not found states, and a very basic meta
sheet.

Further refinement is needed before flagging on, but I would like to
land this progress behind a flag. That way we can flip the flag to allow
for traversal in dev.