Read traversal identity using `ns_sphere_identity` FFI method (#505)
Uses the new FFI method `ns_sphere_identity` introduced in
https://github.com/subconsciousnetwork/noosphere/pull/317 and merged in
noosphere 0.8.4 to reliably retrieve the identity of a sphere
post-traversal.

It's not possible to construct a `Sphere` without knowing its identity
but the result of `ns_traverse_by_petname` gives us a pointer to a
sphere with no identity. This PR makes it possible to construct a
`Sphere` from that pointer without jumping through any hoops.

I was previously using a hack to look up the identity from the address
book but that doesn't work for chained traversal of dotted petnames.