Move all AppDefaults into .start
This makes AppModel testable.

Previously we were loading AppDefaults values directly as default values
when initializing the struct. However, this meant that initialization
was stateful, and the value of those properties depended upon the value
set during the previous test run. Bad!

Now we only load those values into the model when we call `.start`.