Improvements to 3P content view (#508)
Fixes #507 

This PR is pretty minimal. It replaces the read-only editor view with a
new `SubtextView`. This SubtextView is based on ordinary `Text` and
`AttributedString` instances. In future it will allow us to more easily
implement transclusion and other rich features in read-only mode. It
will also be useful for the feed and other story views.

I'd like to keep the PR small, so I am requesting to land this early
version that minimally swaps out the editor for SubtextView, and does
not do much else. It at least provides some visual differentiation vis
editor by making the text `.body` style.