Support multiple peer types (did, petname) in slashlink type (#561)
DIDs are the only stable way to address 3p content, since the petname
for that content could change. To index 3p content, we need to introduce
the concept of an "absolute" slashlink (a slashlink that is rooted to a
did, rather than a petname).

To put it another way, Slashlinks are locators. We also need a stable
identifier.

This PR introduces several new types to resolve the issue:

- `Peer` is an enum that can be either a Did or a Petname. It makes
Slashlink a type that holds a peer and a slug. This concept mirrors the
[Subtext Rust
parser](https://github.com/subconsciousnetwork/subtext/blob/9d61c0461568d1d6168537955a91221ff1ebc34f/parsers/rust/subtext/src/slashlink.rs#L4)
and Noosphere's treatment of Slashlink types.
- `Slashlink` now contains a `Peer` and a `Slug`.
- `Link`, a type containing a `Did` and `Slug`, which provides a stable
identifier for indexing.

We use `Slashlink` wherever we can, and in the few cases where we need
an identifier (database), we resolve the `Slashlink` to a `Link` at
point of use.

This PR also updates the way the database indexes content to use DIDs.

Note this PR deliberately does not address the markup form of
did-slashlinks, only the datatype. For markup form, we are tracking here
https://github.com/subconsciousnetwork/subtext/issues/51

Lays groundwork toward #323, #551, #518

## Limitations

This PR will not correctly re-relativize the slashlinks of 3P content
saved to the database. So if you put something in the database and then
query it again, you'll get a slashlink with a did, instead of a petname.

To get nice petname-slashlinks, we'll need to integrate a new FFI
function to get a petname from a did
https://github.com/subconsciousnetwork/noosphere/issues/358.

However, this PR does not yet index 3P content. It just updates the
system so we can. We should be safe to land it, and then start indexing
3P content once we get the new `DID -> Petname` FFI.