Ensure draft editor loads latest version of its content (#686)
Fixes https://github.com/subconsciousnetwork/subconscious/issues/683

When initially presented, the draft editor has no address associated
with it (because the note lacks a title). Once the note has been saved
(e.g. when leaving the view by tapping a slashlink) it now _has_ an
address. When the draft editor's `.appear` action fired it incorrectly
assumed it had no address and attempts to recreate the from scratch. We
need to check if the `state.address` exists and use _that_ instead of
the stale `description.address`.

h/t @gordonbrander for telling me to check the description logic 🙏 

Simple fix, subtle issue.