Break UserProfileDetailView update functions out
Fixes #734

Believe it or not, this fixes #734. It seems the reason for the
crash is a bug in the Swift compiler. That switch statement, with all
its arms, was just too much for the compiler to handle for some reason.
Breaking out the arms into separate static methods and calling out to
them seems to resolve the issue.

I have some speculation as to why that might be in #734, but really it's
just staring at shadows on cave walls and making up stories.

My takeaway is that we should be breaking functions into bite-sized
chunks so as not to scare the compiler. It tripped up over something
different, but relating to complex Combine types back in #22 #60. It
seems easy to spook.