Make Gateway URL update idempotent (#755)
Related to
https://github.com/subconsciousnetwork/subconscious/issues/693

# Changes
- Clarify gateway URL lifecycle
- Make updates to gateway URL idempotent
- Prevent editing key values during sync

Unsure if this technically "fixes" the issue, since we still reset our
noosphere instance as-soon-as there's a reason to update the gateway
URL. Perhaps we should be queuing the reset operation (somehow) until
after in-flight work is complete? Or perhaps we can find a way to update
the gateway URL without tearing down and rebuilding noosphere?