Work around NavigationStack crash with @State
Replay changes from store on `@State` to work around bug.

Was experiencing a mysterious crasher when factoring NavigationStack
into a subview. What I discovered was that, at some point,
NavigationStack seems to be attempting to mutate the array that
manages that stack. I believe the fact that we pass the stack down
as a value type may be causing the problem. It's obscured because
the issue happens in Apple's proprietary SwiftUI code. The crash
has to do with an array manipulation though, so it is either to do
with the array that manages the view stack, or perhaps it is to do
with some sort of race condition in view rendering vs stack
management on the SwiftUI DOM-patching side.

The workaround was giving NavigationStack a @State binding and
replaying changes onto it.