Massively improve profile load times (#843)
We were _always_ reading the user's sphere to populate the list of notes
on the profile. This PR introduces an optimisation to read notes from
the DB index if we follow the user who's profile we're loading.

Combining this with the existing cache of user profile metadata makes it
practically instant to load any profile we follow.