Extract DetailStackView (#885)
Porting the work from
https://github.com/subconsciousnetwork/subconscious/pull/784, moving
towards https://github.com/subconsciousnetwork/subconscious/issues/641
and https://github.com/subconsciousnetwork/subconscious/issues/322.

This PR extracts the `NavigationStack` logic from
`NotebookNavigationView` and shifts it to the re-usable component
`DetailStackView`. This is a drop-in replacement in terms of the user
experience, a stepping stone towards introducing the profile and feed as
app tabs (next PRs).

I also snuck a couple fixes to warnings in, sorry if the diff is
slightly more confusing as a result.