Add cell subviews to contentView
Unlike normal views, you are not supposed to add subviews directly to
cells. Uncovered via documentation site. This is going to require
reworking some of the constraints. The commit contains partial
constraint fixes for some views.

Additionally, introduced UIViewPreviewRepresentable to allow us to embed
configured UIViews in the canvas for rapid development.