Remove `testWritesThenCloseThenReopenWithVars` (#891)
This test started failing while testing
https://github.com/subconsciousnetwork/noosphere/pull/623, leading us to
question why it ever passed in the first place. Due to Swift compiler
trivia the deinit for `Noosphere` is not guaranteed to be called at the
time we overwrite the variable, meaning two instances of `Noosphere`
co-exist pointing at the same data. We should not ever be trying to open
multiple instances of `Noosphere` pointing at the same database.

After team discussion we decided to drop this test case and make this an
explicit failure case on the Noosphere side.