Fetch basic user metadata from SQLite if available (#925)
Fixes https://github.com/subconsciousnetwork/subconscious/issues/924

This turned into quite a meaty refactor. The goal is to avoid hitting
the network to produce the feed or our local profile view. To accomplish
this we had to revisit our entire approach to loading and caching user
profiles. While a lot of code changed, the new approach is much simpler:
we have two code paths, one for loading user data for list views and
another for the profile detail view.

I've tested loading the feed with several hundred posts in it from
various peers and it loads instantly for me without reading any of the
peer's `_profile_` memos.

# Changes
- Introduce `UserProfileService.identifyUser` which consults the DB to
produce a `UserProfile` and never traverses (intended for list views
which need user metadata)
- If we don't follow this user or their profile is missing from the DB
we fall back to a "shallow" profile display
- Remove `UserProfileService.buildUserProfile`
- Roll `buildUserProfile` into `UserProfileService.loadFullProfileData`
which _always_ fetches the latest profile from noosphere (intended for
the profile detail view)
- We do not fetch profiles for any users in the following list _unless_
we follow them and have their profile indexed in the DB
- Consolidate and fix `UserProfile.aliases` logic
- `EntryStub` now comes with a `did` field

# TODO
- [x] explicitly read from memo for the actual users profile, fall back
to DB
- [x] when listing users if _we_ follow them, check the DB for their
profile, fallback to nothing
- [x] stop passing user profiles around everywhere
- [x] do a join to list the feed?
- [x] show AKA blocks on followed users